Regenerated OCSP response signer certs and put two on the CRL.
To solve a problem in keeping the index.txt file and and CRLs in sync, a
new script, revoke.sh was created.  It gets called after either of the
two repsonse signer certs for Cards #43 and #44 (revoked response signer
certs) are generated.  It also gets called if mkcadata.sh is run to
ensure that the CRL gets updated with the index entries for the two
revoked response signing certs.

Minor bug fix to install-responder.sh where aia/sia/crls were not being
dropped into the right location.